// src/types/IEmployee.ts
export default interface IEmployee {
    firstName: string;
    lastName: string;
    preferredName: string;
    gender: string;
    location: string;
    description: string;
    designation: string;
    workEmail: string;
    workPhone: string;
    linkedinProfile: string;
    profilePicture?: object;
}